Validate External Service Request.
Validates the wallpaper service so that the inputs can be filtered
before we goto the plugin so we can tell the user for valid inputs.
The plugin doesn't through an error if something went wrong, that
needs to be resolved.

As well, refactored the approver service to its own class for
reusability. It will emmit callbacks to the caller when the request
has been denied or succeeded.