Hookup Gdi Encoder windows specific.
Windows XP and 2000 does not recognize JPEG for Active Wallpapers.
It passes but it doesn't set up anything. We would need to convert
everything to BMP. So prepareing project to use GDI instead.

The Codec would be queried from the available formats we require.
This was taken from MSDN docs to get the image codec.