Fix issue when selecting an approval to not affect its data.
When selecting an item in the list, the list element had a change
event causing it to auto save automatically. That is not good
since we are only saving wrt to the buttons remove/removeall.

Since we are touching that code, improve logic so we can reuse
the functionality instead of duplicating.