Preparing DesktopService for Gnome Implementation
Refactored DesktopService into a base class with platform-independent
functionality and a new subclass, WindowsDesktopService, to handle windows
functionality.

Removed functions from DesktopService we don't really use: NewStream() and
StreamDone().

Tweaked how images are converted to BMP. Using IActiveDesktop with jpg files
didn't work on Vista. After some digging it looks like the rules for which
image formats can be used as desktop backgrounds on which version of Windows
using IActiveDesktop are not exactly trivial. So now we always convert to
the lowest common demoninator: BMP. This is what Mozilla does too.